requirements.txt
click==8.1.1
colorama==0.4.4
Flask==2.1.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pip==22.0.4
python-nmap==0.7.1
setuptools==58.1.0
SQLAlchemy==1.4.36
Werkzeug==2.1.0
dash==2.3.1
dash-cytoscape==0.3.0